Refactor: Modularize into distinct components
